Add mi_opaque: a key to find SaveStateEntry
virtio has two DeviceState: one in PCIDevice and the other in VirtIODevice.
Informations such as vmsd is saved in the latter, and the pointer to
the latter is passed to make SaveStateEntry.

There's no good way to get the latter from the PCIDevice pointer.

Therefore, we pass the devstate info when initializing MI cap, so that
we can later find the corresponding dev in SaveStateEntry list.

Signed-off-by: Jintack Lim <jintack@cs.columbia.edu>