[3.0.0 fix] Skip vfio pages in save_zero_page()
Since the commit d7400a3409982a52ac451cd3ca9caee9db670ca7
save_zero_page() is called before ram_save_page(), which probably is
sending vfio pages that we don't want to send. Do the same check in
save_zero_page() as in ram_save_page() to skip the pages.

Note that this is a temporary solution. We got to find a real solution
not to send vfio pages.

Signed-off-by: Jintack Lim <jintack@cs.columbia.edu>

Conflicts:
	migration/ram.c